Create and style the main content
Create index.html and styles/style.css, set basic HTML
template, set the title and add a main element with the
corresponding logo and search form. Configure the body
min-width and min-heigth, centralize the main element,
align and resize the logo and the search form. Add links
to google fonts, customize the search bar and buttons
changing background, borders, font, shadows and the
apperance when hovering or focusing elements.

The objective is to mimic Google page's main elements (the
logo and the search form), restricted to not using developer
tools.