Add and style the footer
Add a footer element with two div, add the Advertising, Business
and How Search works links in the first div and the Privacy,
Terms and Settings link in the second div. Customize the footer
and its links changing height, padding, background color, align
ment, text-decoration and hover effect.

The objective is to mimic Google page's footer, restricted to not
using developer tools.